refactor(progress): rename ProgressBar to TranscriptionProgressCard with new variants
