+++ frontend/src/components/ProgressBar/index.ts
@@ -1,1 +1,1 @@
-export { ProgressBar } from './ProgressBar'
+export { TranscriptionProgressCard } from './TranscriptionProgressCard'
